refactor: Grouping 변경 및 Dependency 추가, 한 파일에 나눠진 타입 분리
