+++ WeatherForecast/WeatherForecast/Extensions/Extension+JSONDecoder.swift
@@ -5,7 +5,7 @@
 //  Created by Marco, Soll on 2021/10/01.
 //
 
-import Foundation
+import Foundation.NSJSONSerialization
 
 extension JSONDecoder {
     convenience init(keyDecodingStrategy: JSONDecoder.KeyDecodingStrategy) {
