+++ WeatherForecast/WeatherForecast/Extensions/Extension+UITableView.swift
@@ -5,7 +5,7 @@
 //  Created by Yongwoo Marco on 2021/10/15.
 //
 
-import UIKit
+import UIKit.UITableView
 
 extension UITableView {
     func initRefresh(targetView: NSObject, action: Selector) {
