+++ WeatherForecast/WeatherForecast/Model/Dependency/WeatherInformation.swift
@@ -1,0 +1,11 @@
+//
+//  WeatherInformation.swift
+//  WeatherForecast
+//
+//  Created by Marco, Soll on 2021/09/28.
+//
+
+protocol WeatherInformation: Codable {
+    var conditions: [Condition] { get }
+    var mainInformation: MainInformation { get }
+}
