+++ WeatherForecast/WeatherForecast/Network/WeatherAPI.swift
@@ -5,7 +5,6 @@
 //  Created by Yongwoo Marco on 2021/10/04.
 //
 
-import Foundation
 import CoreLocation.CLLocation
 
 enum WeatherAPI {
